+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 1.9, 2024-01-26
 ---------------
+- arrange hero spells as a check box grid
 - add remove-buttons to hero army, artifacts and inventory sections
 - ignore swapping blank or otherwise identical slots in inventory and army
 - remember view mode in hero full character sheet dialog